+++ src/hotspot/os/posix/os_posix.hpp
@@ -96,31 +96,6 @@ public:
   static void print_active_locale(outputStream* st);
 };
 
-/*
- * Crash protection for the JfrSampler thread. Wrap the callback
- * with a sigsetjmp and in case of a SIGSEGV/SIGBUS we siglongjmp
- * back.
- * To be able to use this - don't take locks, don't rely on destructors,
- * don't make OS library calls, don't allocate memory, don't print,
- * don't call code that could leave the heap / memory in an inconsistent state,
- * or anything else where we are not in control if we suddenly jump out.
- */
-class ThreadCrashProtection : public StackObj {
-public:
-  static bool is_crash_protected(Thread* thr) {
-    return _crash_protection != NULL && _protected_thread == thr;
-  }
-
-  ThreadCrashProtection();
-  bool call(os::CrashProtectionCallback& cb);
-
-  static void check_crash_protection(int signal, Thread* thread);
-private:
-  static Thread* _protected_thread;
-  static ThreadCrashProtection* _crash_protection;
-  void restore();
-  sigjmp_buf _jmpbuf;
-};
 
 /*
  * This is the platform-specific implementation underpinning
